32-bit compatibility (#999)
* attempt 32-bit compatibility

* don't allow downsampling to Int32: introduces accuracy bugs

* add 0.7 and nightly tests

* allow integers in betweenness centrality

* attempt to fix parallel 32-bit

* work around splitrange issue

* Revert "work around splitrange issue"

This reverts commit 58cbcf8d81f2e771fa19d4c6b73983dac91242fe.

* splitrange overload

+++ src/Parallel/centrality/stress.jl
@@ -1,5 +1,5 @@
 function stress_centrality(g::AbstractGraph,
-    vs::AbstractVector=vertices(g))::Vector{Int}
+    vs::AbstractVector=vertices(g))::Vector{Int64}
 
     n_v = nv(g)
     k = length(vs)
@@ -7,7 +7,7 @@ function stress_centrality(g::AbstractGraph,
 
     # Parallel reduction
     stress = @distributed (+) for s in vs
-        temp_stress = zeros(Int, n_v)
+        temp_stress = zeros(Int64, n_v)
         if degree(g, s) > 0  # this might be 1?
             state = LightGraphs.dijkstra_shortest_paths(g, s; allpaths=true, trackvertices=true)
             LightGraphs._stress_accumulate_basic!(temp_stress, state, g, s)
@@ -19,4 +19,3 @@ end
 
 stress_centrality(g::AbstractGraph, k::Integer) =
     stress_centrality(g, sample(vertices(g), k))
-
